ACP-L1-V2-E03-chat-sequencer: eliminate chat_sessions/internal/service file-count growth
Consolidate the factory-target-catalog fakes/construction/rejection test
files (4 -> 2, no behavior change) instead of registering a larger
package-file-count baseline entry, so this package's PR-owned growth from
sequencing.go/sequencing_test.go nets to zero and the package stays at
origin/main's existing 23-file count rather than growing to 25 under a
newly widened exception. Also updates AdvanceStreamHeadRequest's doc comment
to reflect that Store.AdvanceStreamHead now validates the source identity
tuple against sequencedPositions instead of only carrying it through for
logging.
